requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
coloredlogs==15.0.1
flatbuffers==23.1.4
humanfriendly==10.0
idna==3.4
mpmath==1.2.1
numpy==1.24.1
onnxruntime==1.13.1
onnxruntime-gpu==1.13.1
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
pyreadline3==3.4.1
requests==2.28.2
sympy==1.11.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
typing_extensions==4.4.0
urllib3==1.26.14
